+++ src/prefect/client/client.py
@@ -572,6 +572,10 @@ class Client:
             raise ClientError(
                 "Flows with required parameters can not be scheduled automatically."
             )
+        if any(e.key for e in flow.edges) and flow.result_handler is None:
+            raise ClientError(
+                "Flows are required to have a result handler for storing inputs and outputs."
+            )
         if compressed:
             create_mutation = {
                 "mutation($input: createFlowFromCompressedStringInput!)": {