+++ src/prefect/core/flow.py
@@ -162,9 +162,7 @@ class Flow:
         self.schedule = schedule
         self.environment = environment or prefect.environments.RemoteEnvironment()
         self.storage = storage
-        self.result_handler = (
-            result_handler or prefect.engine.get_default_result_handler_class()()
-        )
+        self.result_handler = result_handler or getattr(storage, "result_handler", None)
 
         self.tasks = set()  # type: Set[Task]
         self.edges = set()  # type: Set[Edge]
@@ -1182,6 +1180,9 @@ class Flow:
 
         Returns:
             - dict representing the flow
+
+        Raises:
+            - ValueError: if `build=True` and the flow has no storage
         """
 
         self.validate()
@@ -1362,6 +1363,10 @@ class Flow:
 
         if labels:
             self.environment.labels.update(labels)
+
+        # register the flow with a default result handler if one not provided
+        if not self.result_handler:
+            self.result_handler = self.storage.result_handler
 
         client = prefect.Client()
         registered_flow = client.register(