+++ src/prefect/engine/task_runner.py
@@ -91,11 +91,7 @@ class TaskRunner(Runner):
     ):
         self.context = prefect.context.to_dict()
         self.task = task
-        self.result_handler = (
-            task.result_handler
-            or result_handler
-            or prefect.engine.get_default_result_handler_class()()
-        )
+        self.result_handler = task.result_handler or result_handler
         super().__init__(state_handlers=state_handlers)
 
     def __repr__(self) -> str: