+++ src/prefect/environments/storage/base.py
@@ -3,6 +3,7 @@ import logging
 from typing import TYPE_CHECKING, Any, List
 
 import prefect
+from prefect.engine.result_handlers import ResultHandler
 from prefect.utilities import logging as prefect_logging
 
 if TYPE_CHECKING:
@@ -12,10 +13,14 @@ if TYPE_CHECKING:
 class Storage(metaclass=ABCMeta):
     """
     Base interface for Storage objects.
+
+    Args:
+        - result_handler (ResultHandler, optional): a default result handler to use for
+            all flows which utilize this storage class
     """
 
-    def __init__(self) -> None:
-        pass
+    def __init__(self, result_handler: ResultHandler = None) -> None:
+        self.result_handler = result_handler
 
     @property
     def labels(self) -> List[str]: